backend/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
alembic==1.17.2
annotated-types==0.7.0
anyio==4.11.0
attrs==25.1.0
Authlib==1.6.6
bcrypt==5.0.0
bidict==0.23.1
certifi==2025.1.31
cffi==2.0.0
charset-normalizer==3.4.2
click==8.3.0
colorama==0.4.6
cryptography==46.0.3
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.119.1
fastapi-cli==0.0.14
fastapi-cloud-cli==0.3.1
frozenlist==1.8.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
multidict==6.7.0
numpy==2.2.3
outcome==1.3.0.post0
packaging==25.0
pandas==2.2.3
passlib==1.7.4
peewee==3.18.2
peewee-migrate==1.13.0
playwright==1.55.0
propcache==0.4.1
psutil==7.0.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pyee==13.0.0
Pygments==2.19.2
PyJWT==2.10.1
pyodbc==5.2.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.13.0
python-multipart==0.0.20
python-socketio==5.16.0
pytz==2025.1
PyYAML==6.0.3
requests==2.32.4
rich==14.2.0
rich-toolkit==0.15.1
rignore==0.7.1
selenium==4.28.1
sentry-sdk==2.42.1
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.44
starlette==0.48.0
tabulate==0.9.0
trio==0.29.0
trio-websocket==0.12.1
typer==0.20.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.38.0
watchfiles==1.1.1
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==15.0.1
wsproto==1.2.0
yarl==1.22.0
